middleware: save exception in variable
79e6565b18cbe31bd478ff3d9af55e98cce59b04 stopped saving the exception
into `e`, which means whenever this gem tries to handle an exception it
throws another due to the use of an "undefined local variable or method
'e'". Fix this by reverting that particular part of the commit, and
continue saving the exception into `e` for use in the log message.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>